+++ lib/screens/home_screen.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:goal_quest/components/custom_clip_path.dart';
+import 'package:goal_quest/components/no_goal_widget.dart';
+import 'package:goal_quest/models/ui_models/goal_card_model.dart';
 import 'package:goal_quest/constants.dart';
 import 'package:goal_quest/styles.dart';
-import 'package:google_fonts/google_fonts.dart';
 
 import 'package:hive_flutter/hive_flutter.dart';
 
@@ -73,7 +74,7 @@ appBar: AppBar(elevation: 0.0,backgroundColor: primaryColor,),
                       'M Y  G O A L S',
                       style: titleFont1,
                     ),
-                    _goalBox.isEmpty ? const NoGoalsText() : const GoalListview()
+                    _goalBox.isEmpty ? const NoGoalsWidget() : const GoalListview()
                   ],
                 ),
               )
@@ -104,24 +105,7 @@ appBar: AppBar(elevation: 0.0,backgroundColor: primaryColor,),
   }
 }
 
-class NoGoalsText extends StatelessWidget {
-  const NoGoalsText({
-    Key? key,
-  }) : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.only(top: 100.0),
-        child: Text(
-          'No goals here yet',
-          style: defaultFont.copyWith(fontStyle: FontStyle.italic),
-        ),
-      ),
-    );
-  }
-}
  // Listview with all goal cards
 class GoalListview extends StatefulWidget {
   const GoalListview({
@@ -145,7 +129,7 @@ class _TestListViewState extends State<GoalListview> {
     );
     createGoalCards() {
       _goalBox.toMap().forEach((key, value) {
-        goalList.add(ReusableGoalCard(
+        goalList.add(GoalCardModel(
           title: value['title'],
           timeSpan: value['timeSpan'],
           creationDate: value['creationDate'],
@@ -159,7 +143,7 @@ class _TestListViewState extends State<GoalListview> {
                       physics: const BouncingScrollPhysics(),
                       children: goalList,
                     )
-                  : const NoGoalsText();
+                  : const NoGoalsWidget();
             });
           }),
         ));
@@ -173,101 +157,5 @@ class _TestListViewState extends State<GoalListview> {
   }
 }
 //A path for the orange container as appbar
-class CustomClipperPath extends CustomClipper<Path> {
-  @override
-  Path getClip(Size size) {
-    double width = size.width;
-    double height = size.height;
-    final path = Path();
 
-    path.lineTo(0, height / 1.5);
-    path.quadraticBezierTo(width * 0.5, height, width, height / 1.5);
-    path.lineTo(width, 0);
-    path.close();
-    return path;
-  }
 
-  @override
-  bool shouldReclip(covariant CustomClipper<Path> oldClipper) {
-    return false;
-  }
-}
-
-class ReusableGoalCard extends HookWidget {
-  final String title;
-  final dynamic timeSpan;
-  final String creationDate;
-  final String dueBeforeDate;
-  final VoidCallback onDelete;
-
-  ReusableGoalCard(
-      {Key? key,
-      required this.title,
-      required this.timeSpan,
-      required this.creationDate,
-      required this.dueBeforeDate,
-      required this.onDelete})
-      : super(key: key);
-  final _goalBox = Hive.box('myGoalBox');
-
-  @override
-  Widget build(BuildContext context) {
-    var isSelected = useState(false);
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 5.0),
-      child: GestureDetector(
-        onTap: () => isSelected.value
-            ? isSelected.value = false
-            : Navigator.pushNamed(context, '/goal_screen', arguments: title),
-        onLongPress: () => isSelected.value = true,
-        child: Card(
-          child: Padding(
-            padding: const EdgeInsets.all(10.0),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                Text(
-                  title,
-                  style: titleFont2,
-                  overflow: TextOverflow.ellipsis,
-                ),
-                const SizedBox(
-                  height: 5,
-                ),
-                Text('time span: $timeSpan days', style: GoogleFonts.dmSans(fontSize: 16)),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Text('Created on: $creationDate', style: subtextFont),
-                    Text(
-                      'Target date: $dueBeforeDate',
-                      style: subtextFont,
-                    )
-                  ],
-                ),
-                Visibility(
-                  visible: isSelected.value,
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.end,
-                    children: [
-                      TextButton.icon(
-                        icon: const Icon(Icons.check_box_outlined),
-                        label: const Text('Mark as done'),
-                        onPressed: (() {}),
-                      ),
-                      const SizedBox(
-                        width: 10,
-                      ),
-                      OutlinedButton(child: const Text('Remove'), onPressed: onDelete),
-                    ],
-                  ),
-                )
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
